Recuperação da 1º versão da classe Produto.java

+++ PastaFontesApp/Produto.java
@@ -1,25 +1,23 @@
-// Classe de objetos que guarda dados de produtos e serviços oferecidos no estacionamento
-// Exemplos de produtos e serviços: estacionar veículo, lavar veículo; trocar óleo
 
 public class Produto {
 
-	private int peso;
+	private double peso;
 	private double altura;
 	
 	public Produto() {
 		
 	}
 	
-	public Produto(int peso, double altura) {
+	public Produto(double peso, double altura) {
 		this.peso=peso;
 		this.altura=altura;
 	}
 	
-	public int getPeso() {
+	public double getPeso() {
 		return peso;
 	}
 	
-	public void setPeso(int peso) {
+	public void setPeso(double peso) {
 		this.peso = peso;
 	}
